support sync formula on isolated-vm execute bundle
