+++ testing/bundle_execution_helper.ts
@@ -1,27 +1,62 @@
-import type {ExecutionContext} from '../api';
+import type {GenericSyncFormula} from '../api';
 import type {PackDefinition} from '../types';
+import type {ParamDefs} from 'api_types';
+import type {ParamValues} from 'api_types';
+import type {SyncExecutionContext} from 'api_types';
+import type {SyncFormulaResult} from '../api';
 import type {TypedStandardFormula} from '../api';
 import  { coerceParams } from './coercion';
 
-export async function executeFormulaWithRawParams(
+export async function executeSyncFormulaWithoutValidation(
+  formula: GenericSyncFormula,
+  params: ParamValues<ParamDefs>,
+  context: SyncExecutionContext,
+  maxIterations: number,
+) {
+  const result = [];
+  let iterations = 1;
+  do {
+    if (iterations > maxIterations) {
+      throw new Error(
+        `Sync is still running after ${maxIterations} iterations, this is likely due to an infinite loop. If more iterations are needed, use the maxIterations option.`,
+      );
+    }
+    let response: SyncFormulaResult<any>;
+    try {
+      response = await formula.execute(params, context);
+    } catch (err) {
+      throw wrapError(err);
+    }
+    result.push(...response.result);
+    context.sync.continuation = response.continuation;
+    iterations++;
+  } while (context.sync.continuation);
+  return result;
+}
+
+
+export async function executeFormulaOrSyncWithRawParams(
   manifest: PackDefinition,
   formulaName: string,
   rawParams: string[],
-  context: ExecutionContext,
+  context: SyncExecutionContext,
 ) {
-  const formula = findFormula(manifest, formulaName);
-  const params = coerceParams(formula, rawParams as any);
-
-  // can't do validation for now due to the dependency of 'url' package (node), see tryParseUrl.
-  // validateParams(formula, params);
-  let result: any;
+  // TODO(huayang): maybe do validating params / results. need to address the url dependency first.
   try {
-    result = await formula.execute(params, context);
+    const formula = tryFindFormula(manifest, formulaName);
+    if (formula) {
+      const params = coerceParams(formula, rawParams as any);
+      return await formula.execute(params, context);
+    }
+
+    const syncFormula = tryFindSyncFormula(manifest, formulaName);
+    if (syncFormula) {
+      const params = coerceParams(syncFormula, rawParams as any);
+      return await executeSyncFormulaWithoutValidation(syncFormula, params, context, 100);
+    }
   } catch (err) {
     throw wrapError(err);
   }
-  // validateResult(formula, result);
-  return result;
 }
 
 export function findFormula(packDef: PackDefinition, formulaNameWithNamespace: string): TypedStandardFormula {
@@ -52,6 +87,38 @@ export function findFormula(packDef: PackDefinition, formulaNameWithNamespace: s
   throw new Error(
     `Pack definition for ${packDef.name} (id ${packDef.id}) has no formula "${name}" in namespace "${namespace}".`,
   );
+}
+
+export function findSyncFormula(packDef: PackDefinition, syncFormulaName: string): GenericSyncFormula {
+  if (!packDef.syncTables) {
+    throw new Error(`Pack definition for ${packDef.name} (id ${packDef.id}) has no sync tables.`);
+  }
+
+  for (const syncTable of packDef.syncTables) {
+    const syncFormula = syncTable.getter;
+    if (syncFormula.name === syncFormulaName) {
+      return syncFormula;
+    }
+  }
+
+  throw new Error(
+    `Pack definition for ${packDef.name} (id ${packDef.id}) has no sync formula "${syncFormulaName}" in its sync tables.`,
+  );
+}
+
+export function tryFindFormula(
+  packDef: PackDefinition, 
+  formulaNameWithNamespace: string,
+): TypedStandardFormula | undefined {
+  try {
+    return findFormula(packDef, formulaNameWithNamespace);
+  } catch (_err) {}
+}
+
+export function tryFindSyncFormula(packDef: PackDefinition, syncFormulaName: string): GenericSyncFormula | undefined {
+  try {
+    return findSyncFormula(packDef, syncFormulaName);
+  } catch (_err) {}
 }
 
 export function wrapError(err: Error): Error {
